Added an arch effect to the game
it is a lot buggy now but theres explosions now. There are also godmode and lot more bullet and more space.

+++ main.py
@@ -17,9 +17,15 @@ from pygame.locals import (
 # Initialize pygame
 pygame.init()
 
+explosions =[]
+
 # Screen dimensions
 SCREEN_WIDTH = 800
-SCREEN_HEIGHT = 400
+SCREEN_HEIGHT = 600
+
+global HADAMART_ON 
+global NORMAL_ON 
+global NOT_ON 
 
 HADAMART_ON = False
 NORMAL_ON = False
@@ -37,9 +43,11 @@ YELLOW = (255, 255, 0)
 GREEN = (0,255,0)
 
 
+global SYSTEM_TEXT 
 SYSTEM_TEXT = "Good Luck :)"
 
-SYSTEM_COLOR = GREEN
+global SYSTEM_COLOR 
+SYSTEM_COLOR  = GREEN
 
 PAUSE = False
 # Define the Player sprite
@@ -82,6 +90,10 @@ class Player(pygame.sprite.Sprite):
 
 # Define the Enemy sprite
 class Enemy(pygame.sprite.Sprite):
+
+    Exploded = False
+    Entangled = False
+    EntangledPartner = None
     def __init__(self):
         super(Enemy, self).__init__()
         self.surf = pygame.Surface((20, 10))
@@ -112,6 +124,43 @@ class Enemy(pygame.sprite.Sprite):
                     self.rect.top = 0
                 if self.rect.bottom >= SCREEN_HEIGHT:
                     self.rect.bottom = SCREEN_HEIGHT
+        for i in entangleds:
+            if self in i:
+                self.Entangled = True
+                index = i.index(self)
+                self.EntangledPartner = i[(index + 1) %2]
+    def explode(self, messagefrom = None):
+        for i in entangleds:
+            if self in i:
+                entangleds.remove(i)
+        self.exploded = True
+        if self.Exploded==True:
+            return
+        print (messagefrom==self, messagefrom==self.EntangledPartner)
+        if not self.Entangled:
+            return
+        if messagefrom == self:
+            return
+        if messagefrom != self.EntangledPartner and messagefrom!=self:
+            self.EntangledPartner.explode(messagefrom = self)
+        for i in enemies:
+            if i== messagefrom or i == self or i==self.EntangledPartner:
+                continue
+
+            if ((i.rect.centerx- self.rect.centerx )**2 + (i.rect.centery-self.rect.centery)**2)**0.5 <=25:
+                 i.explode(messagefrom=self)
+                 i.kill()
+            
+        explosions.append([self.rect.center,15])
+        if ((player.rect.centerx- self.rect.centerx )**2 + (player.rect.centery-self.rect.centery)**2)**0.5 <=25 :
+            gameover(self)
+        if falseplayer != None:
+            if ((falseplayer.rect.centerx- self.rect.centerx )**2 + (falseplayer.rect.centery-self.rect.centery)**2)**0.5 <=25 :
+                gameover(self)
+
+
+            
+        
 
 class Arc(pygame.sprite.Sprite):
     def __init__(self,Enemy1,Enemy2):
@@ -131,6 +180,7 @@ class Arc(pygame.sprite.Sprite):
         if not PAUSE:
             self.rect.move_ip(-self.speed, 0)
             if self.rect.right < 0:
+                self.explode()
                 self.kill()
 
 class Bullet(pygame.sprite.Sprite):
@@ -153,7 +203,6 @@ class Bullet(pygame.sprite.Sprite):
             self.rect.move_ip(self.speed, 0)
             # self.rot_center(45)
             if self.rect.right > SCREEN_WIDTH:
-
                 self.kill()
 
 # Create sprite groups
@@ -177,14 +226,14 @@ all_sprites.add(player)
 # Game variables
 score = 0
 start_time = time.time()
-bullet_count = 10  # Start with 10 bullets
-bullet_limit = 10 #max bullets
+bullet_count = 50  # Start with 10 bullets
+bullet_limit = 50 #max bullets
 font = pygame.font.SysFont('Comic Sans MS', 28)
 
 # Define game events
-CREATING_ENEMY_TIME_INTERVAL = 150
+CREATING_ENEMY_TIME_INTERVAL = 600
 EVENT_TIME_INTERVAL = 4000
-ADD_BULLET_TIME = 5000
+ADD_BULLET_TIME = 2000
 ADDENEMY = pygame.USEREVENT + 1
 DOEVENT = pygame.USEREVENT + 2
 ADDBULLET = pygame.USEREVENT +3
@@ -195,19 +244,54 @@ pygame.time.set_timer(ADDBULLET, ADD_BULLET_TIME)
 
 # Game loop
 clock = pygame.time.Clock()
-running = True
+global running
+running= True
 game_over_message_displayed = False  # Flag for game over message
 falseplayer = None
 
-
 cntr = 0
-while running:
+
+Godmode = False
+
+
+def gameover(enemy):
+
+    global SYSTEM_TEXT
+    global SYSTEM_COLOR
+    global running
+    global game_over_message_displayed
+
+    rng =random.randint(1,2)==1
+    if PAUSE:
+        enemy.explode()
+        enemy.kill()
+
+    elif HADAMART_ON and rng:
+        SYSTEM_TEXT = "Superposition Saved You"
+        SYSTEM_COLOR = GREEN
+        for i in enemies:
+            enemy.explode()
+            i.kill()
+    elif not Godmode:
+        player.kill()
+        running = False
+        game_over_message_displayed = True  # Set the flag
+        game_over_text = font.render("Game Over :(", True, YELLOW)
+
+        screen.blit(game_over_text, (SCREEN_WIDTH // 2 - 100, SCREEN_HEIGHT // 2 - 25))
+    else:
+        SYSTEM_TEXT = "Godmode Saved You"
+        SYSTEM_COLOR = GREEN
+
+while running: # % 25 Hadamart (50 for exit), %25 Not, %30 Entanglement, $20 Nothing
+
 
     if PAUSE:
         cntr += 1
         if cntr>60:
             PAUSE = False
             cntr = 0
+
 
     for event in pygame.event.get():
         if event.type == pygame.QUIT:
@@ -220,9 +304,9 @@ while running:
             bullet_count+=1
         elif event.type == DOEVENT:
 
-            rng = random.randint(0,250)
+            rng = random.randint(0,100)
 
-            if rng<50:
+            if rng<20:
                 SYSTEM_TEXT = ""
                 SYSTEM_COLOR = WHITE
                 PAUSE = False
@@ -231,16 +315,14 @@ while running:
                 PAUSE = True
                 if not HADAMART_ON:
                     
-                    if rng < 100:
+                    if rng < 45:
 
                         SYSTEM_TEXT = "NOT"
                         SYSTEM_COLOR = WHITE
 
                         NOT_ON = not NOT_ON
                         player.rect.center = [player.rect.center[0], SCREEN_HEIGHT-player.rect.center[1]]
-
-
-                    if rng >= 150 and rng <200:
+                    elif rng < 60:
 
                         SYSTEM_TEXT = "HADAMART!!!"
                         SYSTEM_COLOR = RED
@@ -254,7 +336,7 @@ while running:
                         #SYSTEM_TEXT = f"Falseplayer at {falseplayer.rect.centerx} and {falseplayer.rect.centery}"
                     else:
 
-                        System_Text = "Entanglement!"
+                        SYSTEM_TEXT = "Entanglement!"
                         SYSTEM_COLOR = YELLOW
                         breakcount=  0
 
@@ -279,13 +361,14 @@ while running:
 
 
                 else:
-                    HADAMART_ON = False
+                    if rng <=50:
+                        HADAMART_ON = False
 
-                    SYSTEM_TEXT = "Back to Normal"
-                    SYSTEM_COLOR = WHITE
+                        SYSTEM_TEXT = "Back to Normal"
+                        SYSTEM_COLOR = WHITE
 
-                    if falseplayer != None:
-                        falseplayer.kill()
+                        if falseplayer != None:
+                            falseplayer.kill()
 
 
     # Get pressed keys
@@ -297,6 +380,10 @@ while running:
         bullets.add(new_bullet)
         all_sprites.add(new_bullet)
         bullet_count -= 1
+    if HADAMART_ON and pressed_keys[K_SPACE] and falseplayer!= None and bullet_count > 0:
+        new_bullet = Bullet(falseplayer.rect.centerx, player.rect.centery)
+        bullets.add(new_bullet)
+        all_sprites.add(new_bullet)
 
     # Update game objects
     player.update(pressed_keys)
@@ -308,16 +395,23 @@ while running:
 
     pygame.draw.line(screen, (0,255,0), (0,SCREEN_HEIGHT//2), (SCREEN_WIDTH,SCREEN_HEIGHT//2), 3)
 
-
-
     for entity in all_sprites:
         screen.blit(entity.surf, entity.rect)
 
-    entanglelines =[]
+
     for pairs in entangleds:
 
-       a= pygame.draw.line(screen, (225, 255, 0),pairs[1].rect.center, pairs[0].rect.center , 2)
-       entanglelines.append(a)
+       if None in pairs:
+           entangleds.remove(pairs)
+           continue
+       a = pygame.draw.line(screen, (225, 255, 0),pairs[1].rect.center, pairs[0].rect.center , 2)
+
+
+    for i in explosions:
+        if i[1] >0:
+            pygame.draw.circle(screen, (255,0,0), i[0], 25, 25 )
+            i[1]-=1
+
 
     # Display score and timer
     elapsed_time = int(time.time() - start_time)
@@ -335,8 +429,16 @@ while running:
     # Check for collisions
 
     for i in entangleds:
+
+        if None in i:
+            entangleds.remove(i)
+            
+            continue
         if player.rect.clipline((i[0].rect.center), (i[1].rect.center)) != ():
             if PAUSE:
+                entangleds.remove(i)
+                i[0].explode()
+                i[1].explode()
                 i[0].kill()
                 i[1].kill()
 
@@ -344,55 +446,35 @@ while running:
                 SYSTEM_TEXT = "Superposition Saved You"
                 PAUSE = True
                 for i in enemies:
+                    i.explode()
                     i.kill()
                 entangleds.clear()
-            else:
+            elif not Godmode:
                 player.kill()
                 running = False
                 game_over_message_displayed = True  # Set the flag
                 game_over_text = font.render("Game Over :(", True, YELLOW)
 
                 screen.blit(game_over_text, (SCREEN_WIDTH // 2 - 100, SCREEN_HEIGHT // 2 - 25))
+        for bullet in bullets:
+            if bullet.rect.clipline((i[0].rect.center), (i[1].rect.center)) != ():
+                entangleds.remove(i)
+                i[0].explode()
+                i[1].explode()
+                i[0].kill()
+                i[1].kill()
+        
         
 
 
     if pygame.sprite.spritecollideany(player, enemies):
-        enemy = pygame.sprite.spritecollideany(player, enemies)
-
-        if PAUSE:
-            enemy.kill()
-
-        elif HADAMART_ON and random.randint(1,2)==1:
-            SYSTEM_TEXT = "Superposition Saved You"
-            for i in enemies:
-                i.kill()
-        else:
-            player.kill()
-            running = False
-            game_over_message_displayed = True  # Set the flag
-            game_over_text = font.render("Game Over :(", True, YELLOW)
-
-            screen.blit(game_over_text, (SCREEN_WIDTH // 2 - 100, SCREEN_HEIGHT // 2 - 25))
+        gameover(enemy=pygame.sprite.spritecollideany(player, enemies))
     
 
     if falseplayer !=None:
         if pygame.sprite.spritecollideany(falseplayer, enemies):
             enemy = pygame.sprite.spritecollideany(falseplayer, enemies)
-
-            if PAUSE:
-                enemy.kill()
-
-            elif HADAMART_ON and random.randint(1,2)==1:
-                SYSTEM_TEXT = "Superposition Saved You"
-                for i in enemies:
-                    i.kill()
-            else:
-                player.kill()
-                running = False
-                game_over_message_displayed = True  # Set the flag
-                game_over_text = font.render("Game Over :(", True, YELLOW)
-
-                screen.blit(game_over_text, (SCREEN_WIDTH // 2 - 100, SCREEN_HEIGHT // 2 - 25))
+            gameover(enemy=enemy)
 
     for bullet in bullets:
         enemy_hit = pygame.sprite.spritecollideany(bullet, enemies)
@@ -400,6 +482,7 @@ while running:
             for i in entangleds:
                     if enemy_hit in i:
                         entangleds.remove(i)
+            enemy_hit.explode()
             enemy_hit.kill()
             bullet.kill()
             score += 1
